Add process_sample.py.
