+++ score_samples_v2.py
@@ -1,24 +1,58 @@
 """
-Module for processing and scoring dataset samples using xarray and xskillscore.
+Scoring and diagnostics utilities for spatiotemporal ensemble prediction datasets.
 
-Utilities included:
-- Open NetCDF dataset samples (truth and predictions).
-- Compute evaluation metrics (RMSE, MAE, CORR, CRPS, STD_DEV).
-- Extract top-ranked samples based on selected metrics.
-- Flatten and filter NaN values for pointwise exports.
-- Multiprocessing helpers for running over time indices.
+This module provides end-to-end workflows for evaluating model predictions against
+ground truth over time, including:
 
-Notes on performance:
-- `score_samples()` computes full outputs
-  (metrics + error + flattened exports + top samples + p90 grids).
-- `score_samples_multi_ensemble()` is specialized for plotting metrics vs ensembles:
-  it computes *metrics only* for multiple ensemble sizes in a single pass, avoiding
-  flattening/error computation and avoiding re-running the whole pipeline per ensemble.
+- computation of per-time-step skill metrics (RMSE, MAE, CORR, CRPS, etc.)
+- extraction of top-N worst-performing samples for detailed inspection
+- generation of multi-year (e.g. decadal) percentile grids (p90) for selected variables
+- efficient multiprocessing support for large time series
+- optimized handling of ensemble dimensions to avoid redundant I/O and computation
+
+The implementation builds on lower-level helpers from `process_sample` and is
+designed to work with large xarray datasets backed by NetCDF or Dask.
+
+Assumed data layout
+-------------------
+Truth datasets:
+    variables with dims (time, y, x)
+
+Prediction datasets:
+    variables with dims (ensemble, time, y, x)
+    (or already reduced to (time, y, x))
+
+Key design principles
+---------------------
+- Ensemble reductions are explicitly controlled via `n_ensemble`
+- Expensive operations (ensemble mean, file I/O) are minimized and reused
+- All reductions are NaN-aware and spatially consistent
+- Multiprocessing is used at the time-step level for scalability
+- Variable naming is normalized via `VAR_MAPPING` for downstream plotting
+
+Primary entry points
+--------------------
+- score_samples:
+    Compute metrics, spatial errors, top samples, and N-year p90 grids
+    for a single ensemble size.
+
+- score_samples_multi_ensemble:
+    Efficiently compute metrics for multiple ensemble sizes in one pass,
+    reusing loaded data.
+
+Configuration constants
+-----------------------
+- TOP_NUM: number of top samples to extract per metric
+- N_YEARS: window length (years) for p90 aggregation
+- NTH_PERCENTILE: percentile used for 2D grid computation
+- DEBUG: enable optional NetCDF output for p90 grids
+
+This module performs no plotting; all outputs are returned as xarray objects
+for flexible downstream analysis and visualization.
 """
 from __future__ import annotations
 
 import multiprocessing
-import warnings
 from functools import partial
 from typing import Iterable, Dict, List, Tuple, Callable
 
@@ -28,11 +62,11 @@ import pandas as pd
 import xarray as xr
 
 from mask_samples import get_timestamp
+from process_sample import (
+    open_samples, compute_abs_difference,
+    process_sample, process_sample_multi_ensemble,
+)
 
-try:
-    import xskillscore as xs
-except ImportError as exc:
-    raise ImportError("xskillscore not installed. Try `pip install xskillscore`") from exc
 
 DEBUG = False       # Enable to dump p90 netcdf
 TOP_NUM = 5         # Number of top samples to extract
@@ -44,252 +78,6 @@ VAR_MAPPING: Dict[str, str] = {
     "eastward_wind_10m": "u10m",
     "northward_wind_10m": "v10m",
 }
-
-
-# -----------------------------------------------------------------------------
-# IO
-# -----------------------------------------------------------------------------
-def _open_samples(f: str) -> Tuple[xr.Dataset, xr.Dataset, xr.Dataset]:
-    """
-    Open prediction and truth samples from a dataset file.
-
-    Parameters:
-        f: Path to the dataset file.
-
-    Returns:
-        tuple: A tuple containing truth, prediction, and root datasets.
-    """
-    root = xr.open_dataset(f)
-    pred = xr.open_dataset(f, group="prediction").merge(root)
-    truth = xr.open_dataset(f, group="truth").merge(root)
-
-    return truth.set_coords(["lon", "lat"]), pred.set_coords(["lon", "lat"]), root
-
-
-# -----------------------------------------------------------------------------
-# Metrics / transforms
-# -----------------------------------------------------------------------------
-def _compute_crps(truth: xr.Dataset, pred: xr.Dataset) -> xr.Dataset:
-    """
-    Computes the CRPS while filtering out NaN values.
-
-    Parameters:
-        truth (xr.Dataset): The truth dataset.
-        pred (xr.Dataset): The prediction dataset.
-
-    Returns:
-        xr.Dataset: CRPS values computed only for valid points.
-    """
-    dim = ["x", "y"]
-
-    valid_mask = np.isfinite(truth) & pred.notnull().any(dim="ensemble")
-    truth_valid = truth.where(valid_mask)
-    pred_valid = pred.where(valid_mask)
-
-    return xs.crps_ensemble(truth_valid, pred_valid, member_dim="ensemble", dim=dim)
-
-
-def _compute_metrics(truth: xr.Dataset, pred: xr.Dataset) -> xr.Dataset:
-    """
-    Compute RMSE, MAE, CORR, CRPS, and STD_DEV between truth and prediction ensembles.
-
-    Parameters:
-        truth (xr.Dataset): The truth dataset.
-        pred (xr.Dataset): The prediction dataset.
-
-    Returns:
-        xr.Dataset: A dataset containing computed metrics.
-    """
-    dim = ["x", "y"]
-    pred_mean = pred.mean("ensemble")
-
-    rmse = xs.rmse(truth, pred_mean, dim=dim, skipna=True)
-    mae = xs.mae(truth, pred_mean, dim=dim, skipna=True)
-    corr = xs.pearson_r(truth, pred_mean, dim=dim, skipna=True)
-
-    # Compute CRPS and STD_DEV (suppress expected NaN-related warnings)
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore", RuntimeWarning)
-        std_dev = pred.std("ensemble").mean(dim, skipna=True)
-        crps = _compute_crps(truth, pred)
-
-    return (
-        xr.concat([rmse, mae, corr, crps, std_dev], dim="metric")
-        .assign_coords(metric=["RMSE", "MAE", "CORR", "CRPS", "STD_DEV"])
-        .load()
-    )
-
-
-def _flatten_and_filter_nan(truth: xr.Dataset, pred: xr.Dataset) -> Tuple[xr.Dataset, xr.Dataset]:
-    """
-    Extract flattened truth and prediction for all variables in the truth dataset.
-
-    Parameters:
-        truth (xarray.Dataset): The truth dataset.
-        pred (xarray.Dataset): The prediction dataset.
-
-    Returns:
-        tuple: Two xarray.Datasets, one for truth and one for prediction.
-    """
-    truth_data: Dict[str, Tuple[str, np.ndarray]] = {}
-    pred_data: Dict[str, Tuple[str, np.ndarray]] = {}
-
-    for var in truth.data_vars:
-        if var not in pred:
-            continue
-
-        truth_flat = truth[var].values.ravel()
-        pred_flat = pred[var].mean("ensemble").values.ravel() \
-                    if "ensemble" in pred[var].dims else pred[var].values.ravel()
-
-        # Filter out NaNs and align truth and pred
-        valid = np.isfinite(truth_flat) & np.isfinite(pred_flat)
-        truth_data[var] = ("points", truth_flat[valid])
-        pred_data[var] = ("points", pred_flat[valid])
-
-    n_points = len(next(iter(truth_data.values()))[1])
-    coords = {"points": np.arange(n_points)}
-
-    return xr.Dataset(truth_data, coords=coords), xr.Dataset(pred_data, coords=coords)
-
-
-def _compute_abs_difference(truth: xr.Dataset, pred: xr.Dataset) -> xr.Dataset:
-    """
-    Computes the absolute difference between truth and prediction datasets
-    while filtering NaN values.
-
-    Parameters:
-        truth (xr.Dataset): The truth dataset.
-        pred (xr.Dataset): The prediction dataset.
-
-    Returns:
-        xr.Dataset: The absolute difference dataset with NaNs removed.
-    """
-    pred_mean = pred.mean("ensemble").expand_dims("ensemble") \
-                if "ensemble" in pred.dims else pred
-    abs_diff = abs(pred_mean - truth)
-
-    # Filter out NaNs by setting them to NaN where either truth or pred is NaN
-    valid_mask = np.isfinite(truth) & np.isfinite(pred_mean)
-    return abs_diff.where(valid_mask)
-
-
-# -----------------------------------------------------------------------------
-# Multiprocessing helpers
-# -----------------------------------------------------------------------------
-def _run_over_time(n_time: int, worker_fn: Callable[[int], dict],
-                  pool_size: int = 32) -> List[dict]:
-    """Run `worker_fn(time_idx)` over [0..n_time-1] in parallel and collect results."""
-    with multiprocessing.Pool(pool_size) as pool:
-        return list(
-            tqdm.tqdm(
-                pool.imap(worker_fn, range(n_time)),
-                total=n_time,
-            )
-        )
-
-
-def _concat_from_results(results: List[dict], key: str, dim: str) -> xr.Dataset:
-    """Concat `results[i][key]` along `dim` and apply VAR_MAPPING rename."""
-    return xr.concat([res[key] for res in results], dim=dim).rename(VAR_MAPPING)
-
-
-# -----------------------------------------------------------------------------
-# Per-time processing
-# -----------------------------------------------------------------------------
-def _select_time_and_ensemble(
-    filepath: str,
-    index: int,
-    n_ensemble: int
-) -> Tuple[xr.Dataset, xr.Dataset]:
-    """
-    Helper to open samples and select a single time step with optional ensemble slicing.
-    Returns (truth_t, pred_t) loaded into memory.
-
-    Parameters:
-        filepath (str): Path to the dataset file.
-        index (int): The time index to process.
-        n_ensemble (int): Number of ensemble members to consider.
-
-    Returns:
-        Dict[str, xr.Dataset]: A dictionary containing computed metrics, errors,
-        and flattened truth/prediction datasets.
-    """
-    truth, pred, _ = _open_samples(filepath)
-
-    truth_t = truth.isel(time=index).load()
-
-    pred_t = pred.isel(time=index)
-    if n_ensemble > 0 and "ensemble" in pred_t.dims:
-        pred_t = pred_t.isel(ensemble=slice(0, n_ensemble))
-    pred_t = pred_t.load()
-
-    return truth_t, pred_t
-
-
-def _process_sample(index: int, filepath: str, n_ensemble: int) -> Dict[str, xr.Dataset]:
-    """
-    Process a single time step from the dataset.
-
-    Parameters:
-        index (int): The time index to process.
-        filepath (str): Path to the dataset file.
-        n_ensemble (int): Number of ensemble members to consider.
-
-    Returns:
-        Dict[str, xr.Dataset]: A dictionary containing computed metrics, errors,
-        and flattened truth/prediction datasets.
-    """
-    truth_t, pred_t = _select_time_and_ensemble(filepath, index, n_ensemble)
-
-    truth_flat, pred_flat = _flatten_and_filter_nan(truth_t, pred_t)
-
-    return {
-        "metrics": _compute_metrics(truth_t, pred_t),
-        "error": _compute_abs_difference(truth_t, pred_t),
-        "truth_flat": truth_flat,
-        "pred_flat": pred_flat,
-    }
-
-
-def _process_sample_multi_ensemble(
-    index: int,
-    filepath: str,
-    n_ensembles: Tuple[int, ...]
-) -> Dict[int, xr.Dataset]:
-    """
-    Compute metrics for multiple ensemble sizes at a single time index, efficiently.
-
-    Strategy:
-      - open samples once
-      - load truth at time=index once
-      - load prediction at time=index for ensemble[:max_n] once
-      - compute metrics for each n_ens by slicing the already-loaded pred
-
-    Parameters
-    ----------
-    index : int
-        Time index to process.
-    filepath : str
-        Path to NetCDF samples file.
-    n_ensembles : Tuple[int, ...]
-        Ensemble sizes to evaluate (e.g., (1, 4, 16)).
-
-    Returns
-    -------
-    Dict[int, xr.Dataset]
-        Mapping n_ens -> metrics dataset for this single time step.
-    """
-    # Load truth & pred with max ensemble
-    truth_t, pred_t = _select_time_and_ensemble(filepath, index, max(n_ensembles))
-
-    # Compute per-ensemble metrics by slicing in-memory
-    out: Dict[int, xr.Dataset] = {}
-    for n_ens in n_ensembles:
-        pred_n = pred_t.isel(ensemble=slice(0, n_ens)) if "ensemble" in pred_t.dims else pred_t
-        out[n_ens] = _compute_metrics(truth_t, pred_n)
-
-    return {"metrics_by_n": out}
 
 
 # -----------------------------------------------------------------------------
@@ -337,7 +125,7 @@ def _extract_top_samples(
             .load()
         )
 
-        abs_error = _compute_abs_difference(t, p)
+        abs_error = compute_abs_difference(t, p)
         error = abs_error if metric == "MAE" else abs_error ** 2
 
         out[var] = {
@@ -350,7 +138,7 @@ def _extract_top_samples(
 
 
 # -----------------------------------------------------------------------------
-# p90 grids
+# P90 grids
 # -----------------------------------------------------------------------------
 def _window_time_quantile_2d(
     truth: xr.Dataset,
@@ -473,6 +261,25 @@ def p90_by_nyear_period(
 
 
 # -----------------------------------------------------------------------------
+# Multiprocessing helpers
+# -----------------------------------------------------------------------------
+def _run_over_time(n_time: int, worker_fn: Callable[[int], dict],
+                  pool_size: int = 32) -> List[dict]:
+    """Run `worker_fn(time_idx)` over [0..n_time-1] in parallel and collect results."""
+    with multiprocessing.Pool(pool_size) as pool:
+        return list(
+            tqdm.tqdm(
+                pool.imap(worker_fn, range(n_time)),
+                total=n_time,
+            )
+        )
+
+def _concat_from_results(results: List[dict], key: str, dim: str) -> xr.Dataset:
+    """Concat `results[i][key]` along `dim` and apply VAR_MAPPING rename."""
+    return xr.concat([res[key] for res in results], dim=dim).rename(VAR_MAPPING)
+
+
+# -----------------------------------------------------------------------------
 # Scoring entrypoints
 # -----------------------------------------------------------------------------
 def score_samples(
@@ -520,9 +327,9 @@ def score_samples(
     """
     print(f"[{get_timestamp()}] score_samples: filepath={filepath} n_ensemble={n_ensemble}")
 
-    truth, pred, _ = _open_samples(filepath)
+    truth, pred, _ = open_samples(filepath)
     results = _run_over_time(truth.sizes["time"],
-                partial(_process_sample, filepath=filepath, n_ensemble=n_ensemble))
+                partial(process_sample, filepath=filepath, n_ensemble=n_ensemble))
 
     # Metrics
     metrics = _concat_from_results(results, "metrics", dim="time")
@@ -570,9 +377,9 @@ def score_samples_multi_ensemble(
         f"filepath={filepath} n_ensembles={n_ensembles}"
     )
 
-    truth, _, _ = _open_samples(filepath)
+    truth, _, _ = open_samples(filepath)
     results = _run_over_time(truth.sizes["time"],
-                partial(_process_sample_multi_ensemble, filepath=filepath, n_ensembles=n_ensembles))
+                partial(process_sample_multi_ensemble, filepath=filepath, n_ensembles=n_ensembles))
 
     combined: Dict[int, xr.Dataset] = {}
     for n_ens in n_ensembles:
